+++ src/Hacks/LevelFix.cpp
@@ -1,0 +1,15 @@
+#include <Geode/Geode.hpp>
+#include <Geode/modify/LevelPage.hpp>
+#include "../Client/Client.h"
+
+using namespace geode::prelude;
+
+/*class $modify (LevelPage)
+{
+    static LevelPage* create(GJGameLevel* p0)
+    {
+        log::info("level: {}", p0);
+        
+        return LevelPage::create(p0);
+    }
+};*/